+++ lbrynet/core/Wallet.py
@@ -77,7 +77,8 @@ class CachedClaim(object):
         self.height = height
         self.amount = amount
         self.supports = [] if not supports else json.loads(supports)
-        self.effective_amount = self.amount + sum([x['amount'] for x in self.supports])
+        self.effective_amount = self.amount + \
+            sum([x['amount'] for x in self.supports])
         self.channel_name = channal_name
         self.signature_is_valid = signature_is_valid
         self.cache_timestamp = cache_timestamp
@@ -171,7 +172,8 @@ class InMemoryStorage(MetaDataStorage):
             return defer.succeed(None)
 
     def update_claimid(self, claim_id, name, claim_outpoint):
-        self.claimids[(name, claim_outpoint['txid'], claim_outpoint['nout'])] = claim_id
+        self.claimids[(name, claim_outpoint['txid'],
+                       claim_outpoint['nout'])] = claim_id
         return defer.succeed(True)
 
     def get_claimid_for_tx(self, claim_outpoint):
@@ -265,7 +267,7 @@ class SqliteStorage(MetaDataStorage):
         # TODO: refactor the 'claim_ids' table to not be terrible
         txid, nout = claim_outpoint['txid'], claim_outpoint['nout']
         yield self.db.runOperation("INSERT OR REPLACE INTO name_metadata VALUES (?, ?, ?, ?)",
-                                       (name, txid, nout, sd_hash))
+                                   (name, txid, nout, sd_hash))
         defer.returnValue(None)
 
     @rerun_if_locked
@@ -297,7 +299,6 @@ class SqliteStorage(MetaDataStorage):
             response = result[0][0]
         defer.returnValue(response)
 
-
     @rerun_if_locked
     @defer.inlineCallbacks
     def _fix_malformed_supports_amount(self, row_id, supports, amount):
@@ -319,16 +320,16 @@ class SqliteStorage(MetaDataStorage):
             fixed_amount = amount / 100000000.0
             for support in supports:
                 fixed_supports.append(
-                    {'txid':support[0], 'nout':support[1], 'amount':support[2]/100000000.0})
+                    {'txid': support[0], 'nout': support[1], 'amount': support[2] / 100000000.0})
         if fixed_supports is not None:
             log.warn("Malformed support found, fixing it")
             r = yield self.db.runOperation('UPDATE claim_cache SET supports=? WHERE row_id=?',
-                                        (json.dumps(fixed_supports), row_id))
+                                           (json.dumps(fixed_supports), row_id))
             supports = fixed_supports
         if fixed_amount is not None:
             log.warn("Malformed amount found, fixing it")
             r = yield self.db.runOperation('UPDATE claim_cache SET amount=? WHERE row_id=?',
-                                        (fixed_amount, row_id))
+                                           (fixed_amount, row_id))
             amount = fixed_amount
 
         defer.returnValue((json.dumps(supports), amount))
@@ -341,7 +342,8 @@ class SqliteStorage(MetaDataStorage):
                                                "WHERE claimId=?", (claim_id, ))
         response = None
         if r and claim_tx_info and r[0]:
-            rid, _, seq, claim_address, height, amount, supports, raw, chan_name, valid, ts = r[0]
+            rid, _, seq, claim_address, height, amount, supports, raw, chan_name, valid, ts = r[
+                0]
             supports, amount = yield self._fix_malformed_supports_amount(rid, supports, amount)
             last_modified = int(ts)
             name, txid, nout = claim_tx_info[0]
@@ -390,7 +392,7 @@ class SqliteStorage(MetaDataStorage):
                                        "                      certificate_row, last_modified) "
                                        "VALUES (?, ?, ?, ?)",
                                        (uri, cache_row, certificate_row,
-                                       str(int(time.time()))))
+                                        str(int(time.time()))))
         else:
             log.warning("Claim is not in cache")
         defer.returnValue(None)
@@ -435,8 +437,10 @@ class Wallet(object):
         self.wallet_balance = Decimal(0.0)
         self.total_reserved_points = Decimal(0.0)
         self.peer_addresses = {}  # {Peer: string}
-        self.queued_payments = defaultdict(Decimal)  # {address(string): amount(Decimal)}
-        self.expected_balances = defaultdict(Decimal)  # {address(string): amount(Decimal)}
+        # {address(string): amount(Decimal)}
+        self.queued_payments = defaultdict(Decimal)
+        # {address(string): amount(Decimal)}
+        self.expected_balances = defaultdict(Decimal)
         self.current_address_given_to_peer = {}  # {Peer: address(string)}
         # (Peer, address(string), amount(Decimal), time(datetime), count(int),
         # incremental_amount(float))
@@ -451,6 +455,7 @@ class Wallet(object):
 
     def start(self):
         log.info("Starting wallet.")
+
         def start_manage():
             self.stopped = False
             self.manage()
@@ -472,7 +477,8 @@ class Wallet(object):
 
     @staticmethod
     def log_stop_error(err):
-        log.error("An error occurred stopping the wallet: %s", err.getTraceback())
+        log.error("An error occurred stopping the wallet: %s",
+                  err.getTraceback())
 
     def stop(self):
         log.info("Stopping wallet.")
@@ -526,7 +532,8 @@ class Wallet(object):
             def log_error(err):
                 if isinstance(err, AttributeError):
                     log.warning("Failed to get an updated balance")
-                    log.warning("Last balance update: %s", str(self.wallet_balance))
+                    log.warning("Last balance update: %s",
+                                str(self.wallet_balance))
 
             d.addCallbacks(lambda _: self.update_balance(), log_error)
             return d
@@ -535,7 +542,8 @@ class Wallet(object):
 
         def set_next_manage_call():
             if not self.stopped:
-                self.next_manage_call = reactor.callLater(self._balance_refresh_time, self.manage)
+                self.next_manage_call = reactor.callLater(
+                    self._balance_refresh_time, self.manage)
 
         d.addCallback(lambda _: set_next_manage_call())
 
@@ -671,7 +679,8 @@ class Wallet(object):
         payments_to_send = {}
         for address, points in self.queued_payments.items():
             if points > 0:
-                log.debug("Should be sending %s points to %s", str(points), str(address))
+                log.debug("Should be sending %s points to %s",
+                          str(points), str(address))
                 payments_to_send[address] = points
                 self.total_reserved_points -= points
             else:
@@ -680,7 +689,8 @@ class Wallet(object):
             del self.queued_payments[address]
 
         if payments_to_send:
-            log.debug("Creating a transaction with outputs %s", str(payments_to_send))
+            log.debug("Creating a transaction with outputs %s",
+                      str(payments_to_send))
             d = self._do_send_many(payments_to_send)
             d.addCallback(lambda txid: log.debug("Sent transaction %s", txid))
             return d
@@ -749,7 +759,8 @@ class Wallet(object):
                                                         decoded, claim['address'],
                                                         claim['height'],
                                                         claim['amount'], claim['supports'],
-                                                        claim.get('channel_name', None),
+                                                        claim.get(
+                                                            'channel_name', None),
                                                         claim.get('signature_is_valid', None))
         except DecodeError:
             claim['hex'] = claim['value']
@@ -761,7 +772,7 @@ class Wallet(object):
     @defer.inlineCallbacks
     def _handle_claim_result(self, results, update_caches=True):
         if not results:
-            #TODO: cannot determine what name we searched for here
+            # TODO: cannot determine what name we searched for here
             # we should fix lbryum commands that return None
             raise UnknownNameError("")
 
@@ -780,14 +791,14 @@ class Wallet(object):
         # case where return value is {'certificate':{'txid', 'value',...},...}
         if 'certificate' in results:
             results['certificate'] = yield self._decode_and_cache_claim_result(
-                                                                        results['certificate'],
-                                                                        update_caches)
+                results['certificate'],
+                update_caches)
 
         # case where return value is {'claim':{'txid','value',...},...}
         if 'claim' in results:
             results['claim'] = yield self._decode_and_cache_claim_result(
-                                                                     results['claim'],
-                                                                     update_caches)
+                results['claim'],
+                update_caches)
 
         # case where return value is {'txid','value',...}
         # returned by queries that are not name resolve related
@@ -898,7 +909,8 @@ class Wallet(object):
             raise Exception("Invalid channel name")
         elif (parsed_channel_name.path or parsed_channel_name.claim_id or
               parsed_channel_name.bid_position or parsed_channel_name.claim_sequence):
-            raise Exception("New channel claim should have no fields other than name")
+            raise Exception(
+                "New channel claim should have no fields other than name")
         log.info("Preparing to make certificate claim for %s", channel_name)
         return self._claim_certificate(parsed_channel_name.name, amount)
 
@@ -947,7 +959,8 @@ class Wallet(object):
 
         claim = self._process_claim_out(claim)
         claim_outpoint = ClaimOutpoint(claim['txid'], claim['nout'])
-        log.info("Saving metadata for claim %s %d", claim['txid'], claim['nout'])
+        log.info("Saving metadata for claim %s %d",
+                 claim['txid'], claim['nout'])
         yield self._update_claimid(claim['claim_id'], name, claim_outpoint)
         yield self._save_name_metadata(name, claim_outpoint, decoded.source_hash)
         defer.returnValue(claim)
@@ -958,7 +971,7 @@ class Wallet(object):
 
         if not claim_out['success']:
             msg = 'Abandon of {}/{}:{} failed: {}'.format(
-                            claim_id, txid, nout, claim_out['reason'])
+                claim_id, txid, nout, claim_out['reason'])
             raise Exception(msg)
 
         claim_out = self._process_claim_out(claim_out)
@@ -967,7 +980,8 @@ class Wallet(object):
     def support_claim(self, name, claim_id, amount):
         def _parse_support_claim_out(claim_out):
             if not claim_out['success']:
-                msg = 'Support of {}:{} failed: {}'.format(name, claim_id, claim_out['reason'])
+                msg = 'Support of {}:{} failed: {}'.format(
+                    name, claim_id, claim_out['reason'])
                 raise Exception(msg)
             claim_out = self._process_claim_out(claim_out)
             return defer.succeed(claim_out)
@@ -986,7 +1000,8 @@ class Wallet(object):
             result = self._process_claim_out(claim_out)
             defer.returnValue(result)
         else:
-            raise Exception("failed to send tip of %f to claim id %s" % (amount, claim_id))
+            raise Exception(
+                "failed to send tip of %f to claim id %s" % (amount, claim_id))
 
     def get_block_info(self, height):
         d = self._get_blockhash(height)
@@ -1015,12 +1030,14 @@ class Wallet(object):
         balances_to_check = []
         try:
             while self.expected_balance_at_time[0][3] < now:
-                balances_to_check.append(self.expected_balance_at_time.popleft())
+                balances_to_check.append(
+                    self.expected_balance_at_time.popleft())
         except IndexError:
             pass
         ds = []
         for balance_to_check in balances_to_check:
-            log.debug("Checking balance of address %s", str(balance_to_check[1]))
+            log.debug("Checking balance of address %s",
+                      str(balance_to_check[1]))
             d = self._get_balance_for_address(balance_to_check[1])
             d.addCallback(lambda bal: bal >= balance_to_check[2])
             ds.append(d)
@@ -1040,7 +1057,8 @@ class Wallet(object):
                                 balance[4] + 1,
                                 balance[5]
                             )
-                            self.expected_balance_at_time.append(new_expected_balance)
+                            self.expected_balance_at_time.append(
+                                new_expected_balance)
                             peer.update_score(-5.0)
                         else:
                             peer.update_score(-50.0)
@@ -1051,7 +1069,8 @@ class Wallet(object):
                 else:
                     log.warning("Something went wrong checking a balance. Peer: %s, account: %s,"
                                 "expected balance: %s, expected time: %s, count: %s, error: %s",
-                                str(balance[0]), str(balance[1]), str(balance[2]), str(balance[3]),
+                                str(balance[0]), str(balance[1]), str(
+                                    balance[2]), str(balance[3]),
                                 str(balance[4]), str(result.getErrorMessage()))
 
         dl.addCallback(handle_checks)
@@ -1241,7 +1260,8 @@ class LBRYumWallet(Wallet):
         def check_started():
             if self.network.is_connecting():
                 if self._is_first_run():
-                    log.info("Running the wallet for the first time. This may take a moment.")
+                    log.info(
+                        "Running the wallet for the first time. This may take a moment.")
                     self.printed_retrieving_headers = True
                 return False
             self._start_check.stop()
@@ -1249,7 +1269,8 @@ class LBRYumWallet(Wallet):
             if self.network.is_connected():
                 network_start_d.callback(True)
             else:
-                network_start_d.errback(ValueError("Failed to connect to network."))
+                network_start_d.errback(ValueError(
+                    "Failed to connect to network."))
 
         self._start_check = task.LoopingCall(check_started)
 
@@ -1259,7 +1280,8 @@ class LBRYumWallet(Wallet):
         d.addCallback(lambda _: network_start_d)
         d.addCallback(lambda _: self._load_blockchain())
         d.addCallback(lambda _: log.info("Subscribing to addresses"))
-        d.addCallback(lambda _: self.wallet.wait_until_synchronized(lambda _: None))
+        d.addCallback(
+            lambda _: self.wallet.wait_until_synchronized(lambda _: None))
         d.addCallback(lambda _: log.info("Synchronized wallet"))
         d.addCallback(lambda _: self.get_cmd_runner())
         d.addCallbacks(lambda _: log.info("Set up lbryum command runner"))
@@ -1349,7 +1371,7 @@ class LBRYumWallet(Wallet):
         d.addCallback(lambda _: blockchain_caught_d)
         return d
 
-    # run commands as a defer.succeed,
+    # run commands as a defer.succeed ,
     # lbryum commands should be run this way , unless if the command
     # only makes a lbrum server query, use _run_cmd_as_defer_to_thread()
     def _run_cmd_as_defer_succeed(self, command_name, *args, **kwargs):
@@ -1368,13 +1390,23 @@ class LBRYumWallet(Wallet):
         func = getattr(cmd_runner, cmd.name)
         return threads.deferToThread(func, *args, **kwargs)
 
+    @defer.inlineCallbacks
     def _update_balance(self):
         accounts = None
         exclude_claimtrietx = True
-        d = self._run_cmd_as_defer_succeed('getbalance', accounts, exclude_claimtrietx)
-        d.addCallback(
-            lambda result: Decimal(result['confirmed']) + Decimal(result.get('unconfirmed', 0.0)))
-        return d
+        balance = Decimal(0)
+
+        # See issue: https://github.com/lbryio/lbry/issues/626
+        if self.network and self.network.is_connected():
+            balance_info = yield self._run_cmd_as_defer_succeed(
+                'getbalance', accounts, exclude_claimtrietx
+            )
+            balance = Decimal(balance_info['confirmed']) + \
+                      Decimal(balance_info.get('unconfirmed', 0.0))
+        else:
+            log.warning('lbryum is not connected')
+
+        defer.returnValue(balance)
 
     # Always create and return a brand new address
     def get_new_address(self, for_change=False, account=None):
@@ -1447,7 +1479,7 @@ class LBRYumWallet(Wallet):
 
     @defer.inlineCallbacks
     def _send_name_claim(self, name, value, amount,
-                            certificate_id=None, claim_address=None, change_address=None):
+                         certificate_id=None, claim_address=None, change_address=None):
         log.info("Send claim: %s for %s: %s ", name, amount, value)
         claim_out = yield self._run_cmd_as_defer_succeed('claim', name, value, amount,
                                                          certificate_id=certificate_id,
@@ -1476,11 +1508,14 @@ class LBRYumWallet(Wallet):
     def _do_send_many(self, payments_to_send):
         def handle_payto_out(payto_out):
             if not payto_out['success']:
-                raise Exception("Failed payto, reason:{}".format(payto_out['reason']))
+                raise Exception(
+                    "Failed payto, reason:{}".format(payto_out['reason']))
             return payto_out['txid']
 
-        log.debug("Doing send many. payments to send: %s", str(payments_to_send))
-        d = self._run_cmd_as_defer_succeed('payto', payments_to_send.iteritems())
+        log.debug("Doing send many. payments to send: %s",
+                  str(payments_to_send))
+        d = self._run_cmd_as_defer_succeed(
+            'payto', payments_to_send.iteritems())
         d.addCallback(lambda out: handle_payto_out(out))
         return d
 
@@ -1657,7 +1692,8 @@ class LBRYcrdAddressQueryHandler(object):
             d.addCallback(create_response)
             return d
         if self.address is None:
-            log.warning("Expected a request for an address, but did not receive one")
+            log.warning(
+                "Expected a request for an address, but did not receive one")
             return defer.fail(
                 Failure(ValueError("Expected but did not receive an address request")))
         else: