+++ CHANGELOG.md
@@ -22,6 +22,7 @@ at anytime.
   * Fixed handling decryption error for blobs encrypted with an invalid key
   * Fixed handling stream with no data blob (https://github.com/lbryio/lbry/issues/905)
   * Fixed fetching the external ip
+  * Fixed exception when stopping daemon with Ctrl-C (https://github.com/lbryio/lbry/issues/626)
 
 ### Deprecated
   * `channel_list_mine`, replaced with `channel_list`